+++ src/dom/v7400/object.rs
@@ -7,8 +7,8 @@ use crate::dom::error::StructureError;
 use crate::dom::v7400::{Core, Document, DowncastId, NodeId, StrSym, ValidateId};
 use crate::pull_parser::v7400::attribute::DirectAttributeValue;
 
-use self::connection::ConnectionRef;
-pub(crate) use self::graph::ObjectsGraph;
+use self::connection::Connection;
+pub(crate) use self::graph::{ObjectsGraph, ObjectsGraphBuilder};
 pub use self::scene::SceneNodeId;
 
 pub mod connection;
@@ -165,27 +165,18 @@ impl ObjectNodeId {
             .expect("The object node with the `ObjectNodeId` is not stored in the given document")
     }
 
-    /// Returns an iterator of the connections with source objects and
-    /// properties.
+    /// Returns an iterator of the connections.
     ///
-    /// Note that this would not be ordered.
-    /// To access them in correct order, sort by return value of
-    /// [`ConnectionEdge::index()`].
-    pub fn sources_undirected(self, doc: &Document) -> impl Iterator<Item = ConnectionRef<'_>> {
-        self.meta(doc).id().sources_undirected(doc)
+    /// Edges are iterated in same order as raw FBX structure.
+    pub fn sources(self, doc: &Document) -> impl Iterator<Item = &Connection> {
+        self.meta(doc).id().sources(doc)
     }
 
-    /// Returns an iterator of the connections with destination objects and
-    /// properties.
+    /// Returns an iterator of the connections.
     ///
-    /// Note that this would not be ordered.
-    /// To access them in correct order, sort by return value of
-    /// [`ConnectionEdge::index()`].
-    pub fn destinations_undirected(
-        self,
-        doc: &Document,
-    ) -> impl Iterator<Item = ConnectionRef<'_>> {
-        self.meta(doc).id().destinations_undirected(doc)
+    /// Edges are iterated in same order as raw FBX structure.
+    pub fn destinations(self, doc: &Document) -> impl Iterator<Item = &Connection> {
+        self.meta(doc).id().destinations(doc)
     }
 }
 
@@ -217,30 +208,17 @@ impl ObjectId {
         ObjectId(v)
     }
 
-    /// Returns an iterator of the connections with source objects and
-    /// properties.
+    /// Returns an iterator of the connections.
     ///
-    /// Note that this would not be ordered.
-    /// To access them in correct order, sort by return value of
-    /// [`ConnectionEdge::index()`].
-    pub fn sources_undirected(self, doc: &Document) -> impl Iterator<Item = ConnectionRef<'_>> {
-        doc.objects_graph()
-            .incoming_edges_unordered(self)
-            .map(|(src, dest, edge)| ConnectionRef::new(src, dest, edge))
+    /// Edges are iterated in same order as raw FBX structure.
+    pub fn sources(self, doc: &Document) -> impl Iterator<Item = &Connection> {
+        doc.objects_graph().incoming_edges(self)
     }
 
-    /// Returns an iterator of the connections with destination objects and
-    /// properties.
+    /// Returns an iterator of the connections.
     ///
-    /// Note that this would not be ordered.
-    /// To access them in correct order, sort by return value of
-    /// [`ConnectionEdge::index()`].
-    pub fn destinations_undirected(
-        self,
-        doc: &Document,
-    ) -> impl Iterator<Item = ConnectionRef<'_>> {
-        doc.objects_graph()
-            .outgoing_edges_unordered(self)
-            .map(|(src, dest, edge)| ConnectionRef::new(src, dest, edge))
+    /// Edges are iterated in same order as raw FBX structure.
+    pub fn destinations(self, doc: &Document) -> impl Iterator<Item = &Connection> {
+        doc.objects_graph().outgoing_edges(self)
     }
 }